Add initial conflict resolver for deduper.
This first conflict resolver allows fields to be specified that are booleans & should be resolved to TRUE

A few thoughts
1) I looked at how to align with the resolvers in https://github.com/systopia/de.systopia.xdedupe
- a fundamental difference, however, is that Bjoern's work conceptually involves resolving
conflicts and then calling the Contact.merge action. Our logic to date has been to rely on
'other things' (batch job, deduper, legacy form) calling the Merge action and then
to resolve conflicts via a hook. I don't see us switching as we currently have a scenario where we can
hook into Bjoern's screen or any other dedupe methodology, whereas his resolvers only work from his flow.
That makes it a little different to figure out how to integrate the two but I'm still pondering.

2) The resolution in the Deduper is site specific - ie ALWAYS resolve in this way in Batch
mode - which means the batch script, the deduper, any api that calls Conatct.merge but not the
inbuilt user form.

3) I need to tackle 'on_hold' for email - I suspect that is not resolving yet & potentially
add a test for that and the custom_data fields.