Fix type issue when merging location type
This addresses an issue where we are merging a contact with
- email address = home (primary)
- email address = billing
- email address = main

Into a contact with
- email address = home (primary)
- email address = billing

Where only the 2 home primary addresses are the same.

Note it is important here that main has a lower location type ID than
billing does. When the merge takes place the location arrays are keyed by
location type ID - meaing that in the contact to remove
the block number for the billing location is 2
whereas for the contact to keep it is 1 - the code is expecting them
to have the same block number and a recent-ish addition
of typing made a soft miss into a hard error

Store full details of location conflicts on class

This adds new helpers to get the details of the conflicts, complete
with the associated entities.

This is needed because the code was assuming looking for conflicts by block number. The block
number on the 'other' contact is 'significant' in that it is what the conflict is
keyed by - but it turns out it is in conflict with address with the same location on
the main conflict - which may, or may not have the same block number