Store full details of location conflicts on class
This adds new helpers to get the details of the conflicts, complete
with the associated entities.

This is needed because the code was assuming looking for conflicts by block number. The block
number on the 'other' contact is 'significant' in that it is what the conflict is
keyed by - but it turns out it is in conflict with address with the same location on
the main conflict - which may, or may not have the same block number